Module type Signed_route.S

module Request : Request.S

module Response : Response.S

val method_name : string

Human readable method name.

module Signer : sig ... end

module Error : sig ... end

val http_request : Signer.t -> Request.t -> Http_types.Request.t

http_request req constructs a signed http request for Request.t.

val dispatch : ?timeout:Core.Time_ns.Span.t -> Http_client.t -> Signer.t -> Request.t -> (Response.t, Error.t) Core.Result.t Async.Deferred.t

dispatch ?timeout client signer req constructs a signed http request and dispatches using Http_client.t.

val dispatch_exn : ?here:Stdlib.Lexing.position -> ?timeout:Core.Time_ns.Span.t -> Http_client.t -> Signer.t -> Request.t -> Response.t Async.Deferred.t

dispatch_exn ?timeout client signer req constructs a signed http request and dispatches using Http_client.t.