Module Http_types.Query

module Param : sig ... end

type t = Param.t list

The `t` type represents a list of query parameters as a list of key-value pairs.

val empty : t

empty empty list of query parameters.

val with_param : t -> key:string -> value:string -> t

with_param q ~key ~value creates a new list of query parameters by adding a new key-value pair with the given key key and value value to the list of query parameters q.

val sort : t -> t

sort q sorts the given list of query parameters q by the keys.