Module Request_id

include Core.Identifiable

type t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val of_string : string -> t

val to_string : t -> string

val pp : Base__.Formatter.t -> t -> unit

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val validate_lbound : min:t Core__.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core__.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core__.Maybe_bound.t -> max:t Core__.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : sig ... end

val comparator : ( t, comparator_witness ) Core__Comparator.comparator

module Map : sig ... end

module Set : sig ... end

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val hashable : t Core__.Hashtbl.Hashable.t

module Table : sig ... end

module Hash_set : sig ... end

module Hash_queue : sig ... end

val create : unit -> ?now:Core.Time_ns.Span.t -> unit -> t

Creates a function for generating new request IDs and starting new requests. The generated requests will have sequential IDs starting from 0.

val finish_ : ?now:Core.Time_ns.Span.t -> t -> t

Finishes the given request and returns the finished request. This function expects a Started request, and will raise if called on a Finished request.

val id : t -> int

Returns unique ID of the request.

val span : t -> Core.Time_ns.Span.t option

Returns the duration of the given request as a Time_ns.Span.t, or None if the request is not finished.