Module Route.Make

Parameters

module E : Descriptor

Signature

module Request : Request.S with type t = E.Request.t

module Response : Response.S with type t = E.Response.t with type Error.t = E.Response.Error.t

val method_name : string

Human readable method name.

module Error : sig ... end

val http_request : Request.t -> Http_types.Request.t

http_request req constructs http request for Request.t.

val dispatch : ?timeout:Core.Time_ns.Span.t -> Http_client.t -> Request.t -> (Response.t, Error.t) Core.Result.t Async.Deferred.t

dispatch ?timeout client req constructs http request and dispatches using Http_client.t.

val dispatch_exn : ?here:Stdlib.Lexing.position -> ?timeout:Core.Time_ns.Span.t -> Http_client.t -> Request.t -> Response.t Async.Deferred.t

dispatch_exn ?timeout client req constructs http request and dispatches using Http_client.t.